Updates to search functionality
App can now search for all tickers available with the MarketStack API. When a user starts typing a list of suggestions pop down and when a user clicks a suggestion the marketCard for the appropriate stock is shown. Users can search by stock name or symbol. Users can return to the dashboard view once done searching